erasure-code: jerasure preloads all variants
All jerasure plugin are preloaded, regardless of the CPU features
available. The selection of the variant happens in the factory() method
and will not use a variant that would require CPU features that are not
available.

Although it would be possible to only load the plugins that will be
used, it would not allow the caller to override the selection to use the
generic version suitable for all processors over the SSE4 version that
would be automatically selected or even the SSE3 version. Such an
override is useful when running benchmarks, to compare the performances
of the variants.

http://tracker.ceph.com/issues/9153 Fixes: #9153

Signed-off-by: Loic Dachary <loic-201408@dachary.org>